enclave/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enclave"
version = "0.0.1"
requires-python = ">=3.9"
dependencies = [
  "core",
  "pydantic==2.10.3",
  "uvicorn==0.34.0",
  "fastapi==0.115.6",
  "cryptography==42.0.5",
  "click==8.1.8",
  "requests==2.32.3",
  "web3==7.8.0",
  "eth-account==0.13.5",
  "python-dotenv==1.0.1",
  "openai==1.65.5",
]

[project.optional-dependencies]
dev = ["mypy==1.15.0"]

[project.scripts]
enclave = "enclave.main:main"
host_to_vsock = "enclave.host_to_vsock:main"
host_to_remote = "enclave.host_to_remote:main"
